Separate class files for User implementation
