+++ LDAP/app/package.json
@@ -4,7 +4,7 @@
     "prebuild": "tslint -c tslint.json -p tsconfig.json --fix",
     "build": "tsc",
     "prestart": "npm run build",
-    "start": "node dist/index.js",
+    "start": "node dist/main.js",
     "start2": "node dist/users.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
@@ -14,5 +14,8 @@
     "ldapjs": "^1.0.2",
     "tslint": "^6.0.0",
     "typescript": "^3.8.3"
+  },
+  "devDependencies": {
+    "@types/ldapjs": "^1.0.7"
   }
 }
