+++ LDAP/app/src/LdapTypes/CommonName.ts
@@ -1,0 +1,12 @@
+import { default as LdapKeyValuePair } from "./LdapKeyValuePair";
+
+export default class CommonName extends LdapKeyValuePair
+    {
+        public constructor(cn:string)
+        {
+            super("cn",cn);
+        }
+        public toString():string{
+            return this.value;
+        }
+    }
