+++ LDAP/app/src/LdapTypes/LdapKeyValuePairBag.ts
@@ -1,0 +1,13 @@
+import { default as LdapKeyValuePair } from "./LdapKeyValuePair";
+
+export default class LdapKeyValuePairBag
+    {
+        protected bag:LdapKeyValuePair[];
+        protected constructor(bag:LdapKeyValuePair[])
+        {
+            this.bag=bag;
+        }
+        public toArray():LdapKeyValuePair[]{
+            return this.bag;
+        }
+    }
