+++ LDAP/app/src/LdapTypes/UserID.ts
@@ -1,0 +1,11 @@
+import { default as LdapKeyValuePair } from "./LdapKeyValuePair";
+export default class UserID extends LdapKeyValuePair
+    {
+        public constructor(uid:string)
+        {
+            super("uid",uid);
+        }
+        public toString():string{
+            return this.value;
+        }
+    }
