+++ LDAP/app/src/main.ts
@@ -1,0 +1,5 @@
+// import "users.ts";
+import { User } from "./users";
+// let users:any = require("users.ts");
+
+User.loadUser("uid=dummyaccount,ou=people,dc=linuxlab,dc=salisbury,dc=edu");
